nromalise percentiles

+++ data_collection/get_percentiles.py
@@ -16,37 +16,45 @@ mode = args.data
 fits_dir = f"DATA/fits_{mode}/"
 np_dir = f"DATA/np_{mode}/"
 os.makedirs(np_dir) if not os.path.exists(np_dir) else None
-files = os.listdir(fits_dir)
+already_done = os.listdir(np_dir)
+files = np.sort(os.listdir(fits_dir))
 dates = []
 q = [0, 0.01, 0.1, 1, 5, 10, 25, 50, 75, 90, 95, 99, 99.9, 99.99, 100]
 
 
 def get_percentiles(filename):
     try:
-        print(filename)
-        map_ref = sunpy.map.Map(fits_dir + filename)
+        name = filename.strip(".fits").strip('.fts')
+        if name + ".npy" in already_done:
+            print(f"Already converted: {name}")
+            data = np.load(f"{np_dir}{name}.npy").flatten()
+        else:
+            print(filename)
+            map_ref = sunpy.map.Map(fits_dir + filename)
 
-        # not necessary for percentiles
-        mat = map_ref.rotation_matrix
-        map_ref = map_ref.rotate(rmatrix=mat)
+            # not necessary for percentiles
+            mat = map_ref.rotation_matrix
+            map_ref = map_ref.rotate(rmatrix=mat)
 
-        # crop so only sun is shown
-        radius = map_ref.rsun_obs
-        top_right = SkyCoord(radius, radius,
-                             frame=map_ref.coordinate_frame)
-        bottom_left = SkyCoord(-radius, -radius,
-                               frame=map_ref.coordinate_frame)
-        submap = map_ref.submap(bottom_left, top_right)
+            # crop so only sun is shown
+            radius = map_ref.rsun_obs
+            top_right = SkyCoord(radius, radius,
+                                 frame=map_ref.coordinate_frame)
+            bottom_left = SkyCoord(-radius, -radius,
+                                   frame=map_ref.coordinate_frame)
+            submap = map_ref.submap(bottom_left, top_right)
 
-        data = submap.data
-        if data is not None:
-            name = filename.strip(".fits").strip('.fts')
-            np.save(f"{np_dir}{name}", data)
-            data = np.nan_to_num(data).flatten()
-            percentiles = np.percentile(data, q)
-            if percentiles is not None:
-                append_date(name)
-                return percentiles
+            data = submap.data
+            if data is not None:
+                np.save(f"{np_dir}{name}", data)
+                data = np.nan_to_num(data).flatten()
+            else:
+                return
+
+        percentiles = np.percentile(data, q)
+        if percentiles is not None:
+            append_date(name)
+            return percentiles
     except TypeError as err:
         print(f"TypeError:{filename}, {err}")
     except OSError as err:
@@ -55,6 +63,7 @@ def get_percentiles(filename):
         print(f"ValueError:{filename}, {err}")
     except IndexError as err:
         print(f"IndexError:{filename}, {err}")
+    return
 
 
 def append_date(name):
